+++ oarepo_model_builder/property_preprocessors/text_keyword.py
@@ -1,6 +1,6 @@
 from oarepo_model_builder.builders.jsonschema import JSONSchemaBuilder
 from oarepo_model_builder.builders.mapping import MappingBuilder
-from oarepo_model_builder.preprocessors import PropertyPreprocessor, process
+from oarepo_model_builder.property_preprocessors import PropertyPreprocessor, process
 from oarepo_model_builder.schema import deepmerge
 from oarepo_model_builder.stack import ModelBuilderStack
